refactor/camera-screen-component: use index file for camera screen components
