+++ src/CameraScreen/components/index.ts
@@ -1,0 +1,4 @@
+export {default as Controls} from "./Controls";
+export * from "./DeviceNotFound";
+export {default as Hint} from "./Hint";
+export * from "./NoPermissionsPlaceholder";
